fix logs
